Added TTS class
Created and implemented a TTS class, as well as a testing class for it.
We can use the SpeechDriver differently, however for now it's simply
TTS. I dont like the way it handles creating a new object for speaking,
but that's something that can easily change.

Voice needs to change.