reverted queue system in TTS
in testing, it appears that the FreeTTS software has a built in queue
system for the TTS inputs. If that's the case, I don't need my own
queue. That being said, I commented out the queue lines of code, and if
it turns out we need them we can add them later (or revert this commit).