Adding structure to allow for patching, however
I've realized that I have a huge flaw with my current packet setup. It
works perfectly for non-communication commands, such as Shutdown and
Command and Notification. However, it fails largely when I want to do
something like "Do you want to patch?" "Yes" "Patching", as the packets
are each interpretted as if they're their own conversation. So, will now
be branching to solve this issue first, then coming back to patches