Shard_Core framework built for patching
I finally got the shard core to wait for patching to continue before
allowing actions. It also will allow the patching of files. The
ShardPatcher class has a lot more helper methods, need to add
documentation to them however. Currently, it checks if the versions are
the same. If not, it calls upon the downloading of the patch file (does
nothing yet), and then calls to run the patch file (does nothing yet).